+++ pyat/at/lattice/__init__.py
@@ -4,8 +4,7 @@ Helper functions for working with AT lattices.
 A lattice as understood by pyAT is any sequence of elements.  These functions
 are useful for working with these sequences.
 """
-print('lattice start')
-from .utils import *
+
 from .elements import *
+from .utils import *
 from .lattice_object import *
-print('lattice end')
