+++ pyat/at/load/__init__.py
@@ -2,6 +2,5 @@
 Import of AT lattice from different formats:
 - .mat files
 """
-print('load start')
-from at.load.matfile import *
-print('load end')
+
+from .matfile import *
