+++ pyat/at/physics/__init__.py
@@ -2,12 +2,10 @@
 Accelerator physics functions
 """
 
-print('physics start')
 from .orbit import *
+from .amat import *
 from .matrix import *
 from .linear import *
 # noinspection PyUnresolvedReferences
 from .diffmatrix import find_mpole_raddiff_matrix
-from .amat import *
 from .radiation import *
-print('physics end')
