+++ pyat/at/physics/linear.py
@@ -3,9 +3,9 @@ Coupled or non-coupled 4x4 linear motion
 """
 import numpy
 from math import sqrt, atan2, pi
-from ..lattice import Lattice, uint32_refpts, get_s_pos, AtError
-from ..tracking import lattice_pass
-from ..physics import find_orbit4, find_m44, jmat
+from at.lattice import Lattice, uint32_refpts, get_s_pos, AtError
+from at.tracking import lattice_pass
+from at.physics import find_orbit4, find_m44, jmat
 
 __all__ = ['get_twiss', 'linopt', 'get_mcf']
