+++ pyat/at/physics/matrix.py
@@ -5,33 +5,14 @@ A collection of functions to compute 4x4 and 6x6 transfer matrices
 """
 
 import numpy
-from scipy.linalg import block_diag
 from at.lattice import Lattice, uint32_refpts
 from at.tracking import lattice_pass, element_pass
-from at.physics import find_orbit4, find_orbit6
+from at.physics import find_orbit4, find_orbit6, jmat
 
-__all__ = ['find_m44', 'find_m66', 'find_elem_m66', 'jmat']
+__all__ = ['find_m44', 'find_m66', 'find_elem_m66']
 
 XYDEFSTEP = 6.055454452393343e-006  # Optimal delta?
 DPSTEP = 6.055454452393343e-006  # Optimal delta?
-
-# Prepare symplectic identity matrix
-_j2 = numpy.array([[0., 1.], [-1., 0.]])
-_jm = [_j2, block_diag(_j2, _j2), block_diag(_j2, _j2, _j2)]
-
-
-def jmat(ind):
-    """
-    Return the antisymetric block diagonal matrix [[0, 1][-1, 0]]
-
-    INPUT
-        ind     1, 2 or 3. Matrix dimension
-
-    OUTPUT
-        jm      block diagonal matrix, (2, 2) or (4, 4) or (6, 6)
-    """
-    return _jm[ind - 1]
-
 
 _jmt = jmat(2)
