+++ pyat/at/tracking/__init__.py
@@ -2,9 +2,7 @@
 Tracking functions
 """
 
-print('tracking init')
-from .patpass import patpass
-from .track import *
 # noinspection PyUnresolvedReferences
 from .atpass import atpass, elempass
-print('tracking end')
+from .patpass import patpass
+from .track import *
